rearrange folder and encapsulate spotifyRoute.ts
