+++ backend/Node/routes/openWeatherRoute.ts
@@ -3,7 +3,7 @@ import { reverseWeather } from './route-helpers/openweather-helpers';
 
 const router = express.Router();
 
-router.get('/', (req, res) => {
+router.get('/', (_, res) => {
 	res.send('Open Weather API');
 });
