+++ backend/Node/routes/route-helpers/spotify-helpers/spotify-mapper.ts
